feat: 버전 얻어오는 로직을 커스텀 훅으로 분리
