+++ packages/web/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useMemoDeleteMutation } from './useMemoDeleteMutation';
 export { default as useLanguage } from './useLanguage';
+export { default as useGetExtensionVersion } from './useGetExtensionVersion';
